python-services/rag_service/requirements.txt
python-socketio==5.10.0
aiohttp==3.9.1
pymongo==4.6.1
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
faiss-cpu==1.7.4
openai>=1.10.0,<2.0.0
python-dotenv==1.0.0 